requirements.txt
streamlit>=1.38
google-genai>=0.1.0
Pillow